@@ -1,18 +1,22 @@
 {
   "name": "@dontcode/sandbox",
-  "version": "0.5.41",
+  "version": "0.5.42-1",
   "peerDependencies": {
-    "@angular/common": "^13.3.2",
-    "@angular/core": "^13.3.2",
-    "@angular/router": "^13.3.2",
-    "primeflex": "^3.1.3",
-    "primeicons": "^5.0.0",
-    "primeng": "^13.3.2",
-    "@dontcode/core": "^0.5.41",
-    "@dontcode/plugin-common": "^0.5.41",
-    "rxjs": "^7.5.5"
+    "@angular/common": "^15.1.1",
+    "@angular/core": "^15.1.1",
+    "@angular/router": "^15.1.1",
+    "primeflex": "^3.3.0",
+    "primeicons": "^6.0.1",
+    "primeng": "^15.1.0",
+    "@dontcode/core": "^0.6.1",
+    "@dontcode/plugin-common": "^0.6.1",
+    "rxjs": "^7.8.0",
+    "async-mutex": "^0.4.0",
+    "broadcast-channel": "^4.20.2",
+    "core-js": "^3.27.2",
+    "dexie": "^3.2.2"
   },
   "dependencies": {
-    "tslib": "^2.3.1"
+    "tslib": "^2.4.1"
   }
 }
