@@ -1,5 +1,7 @@
-module.exports = {
+/* eslint-disable */
+export default {
   displayName: 'common',
+  preset: '../../jest.preset.js',
 
   setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
   globals: {
@@ -17,6 +19,5 @@ module.exports = {
     'jest-preset-angular/build/serializers/no-ng-attributes',
     'jest-preset-angular/build/serializers/ng-snapshot',
     'jest-preset-angular/build/serializers/html-comment',
-  ],
-  preset: '../../jest.preset.ts',
+  ]
 };
